Move BFGS library
