+++ include/LBFGS/LineSearchBracketing.h
