+++ src/distgeom.cpp
@@ -26,8 +26,8 @@ GNU General Public License for more details.
 #include <openbabel/builder.h>
 #include <openbabel/elements.h>
 #include <openbabel/generic.h>
-#include <openbabel/LBFGS.h>
 #include "rand.h"
+#include <LBFGS.h>
 
 #include <openbabel/stereo/stereo.h>
 #include <openbabel/stereo/cistrans.h>
